update app category update search poi (not done) update detail poi (not done)
